requirements.txt
asgiref==3.5.2
Django
dj-database-url==0.5.0
django-cors-headers==3.10.1
djangorestframework==3.13.1
gunicorn==20.1.0
imap-tools==0.56.0
Jinja2==3.0.3
pillow
psycopg2-binary==2.9.3
psycopg2==2.7.7
pymongo
python-decouple==3.5
requests==2.27.1
xmltodict==0.12.0

